MOD: checkpoints methods

+++ src/subcontracts/PLMData.sol
@@ -124,6 +124,6 @@ contract PLMData is IPLMData {
         pure
         returns (uint256)
     {
-        return charInfo.level**3;
+        return uint256(charInfo.level)**3;
     }
 }
